Prevent writing to changelog in dry run

+++ index.js
@@ -54,11 +54,14 @@ class KeepAChangelog extends Plugin {
   }
 
   beforeRelease() {
-    const { version } = this.getContext();
-    const formattedDate = getFormattedDate();
-    const releaseTitle = `\n\n## [${version}] - ${formattedDate}\n\n`;
-    const changelog = this.changelogContent.replace(this.unreleasedTitle, releaseTitle);
-    fs.writeFileSync(this.changelogPath, changelog);
+    const { isDryRun } = this.global;
+    if (!isDryRun) {
+      const { version } = this.getContext();
+      const formattedDate = getFormattedDate();
+      const releaseTitle = `\n\n## [${version}] - ${formattedDate}\n\n`;
+      const changelog = this.changelogContent.replace(this.unreleasedTitle, releaseTitle);  
+      fs.writeFileSync(this.changelogPath, changelog);
+    }
   }
 }
